Rename and redistribute utils to specific utilities modules
