+++ cdptools/legistar_utils/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Legistar utilities package for cdptools."""
