+++ cdptools/pipelines/event_pipeline.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 from .. import get_module_version
-from ..utils import RunManager
+from ..dev_utils import RunManager
 from .pipeline import Pipeline
 
 ###############################################################################
