+++ cdptools/research_utils/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""Research utilities package for cdptools."""
+
+from .transcripts import download_most_recent_transcripts  # noqa: F401
