+++ cdptools/tests/dev_utils/test_run_manager.py
@@ -9,8 +9,8 @@ from firebase_admin import firestore
 from google.cloud import storage
 
 from cdptools.databases.cloud_firestore_database import CloudFirestoreDatabase
+from cdptools.dev_utils import RunIO, RunManager
 from cdptools.file_stores.gcs_file_store import GCSFileStore
-from cdptools.utils import RunIO, RunManager
 
 from ..databases.test_cloud_firestore_database import MockedCollection
 from ..file_stores.test_gcs_file_store import MockedBlob, MockedBucket
