+++ cdptools/tests/research_utils/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Unit test package for cdptools.research_utils."""
