--- cdptools/tests/utils/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Unit test package for cdptools.utils."""
