--- cdptools/utils/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Utils package for cdptools."""
-
-from .research_utils import download_most_recent_transcripts  # noqa: F401
-from .run_manager import RunIO, RunManager  # noqa: F401
